good automation, try to extend to list resa

+++ templates/delete.sh
@@ -1,13 +1,22 @@
-# cleanup the stacks
-aws cloudformation delete-stack --stack-name skiagenda-website
+LAYER=$1
+if [ -z $LAYER ]
+then
+    LAYER = 'backend'
+fi
+
+# cleanup the stack
+aws cloudformation delete-stack --stack-name skiagenda-$LAYER
 
 # wait for completion
 aws cloudformation wait stack-delete-complete \
-    --stack-name skiagenda-website
+    --stack-name skiagenda-$LAYER
 
-# empty the website bucket
-aws s3 rb s3://skiagenda --force
+if [ $LAYER = 'website' ]
+then
+    # empty the website bucket
+    aws s3 rb s3://skiagenda --force
 
-# cleanup the temporary S3 bucket
-aws s3 rb s3://skiagenda-source --force
+    # cleanup the temporary S3 bucket
+    aws s3 rb s3://skiagenda-source --force
+fi
